fix benchmarks
